scraper/requirements.txt
SportsCardTool==0.2.4
pandas==2.2.2
python-dotenv==1.0.1
mysql-connector-python==8.3.0
loguru==0.7.2

